+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add packagemanager#has_package
 * Add FRESH_CONNECT to all requests to ensure no cached connection is used
 * Add sling#is_valid_login
+* Add debug field to PyAemResult
 
 ### 0.9.1
 * Add file_path arg to webconsole#install_bundle to allow custom bundle file location